requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
twilio==8.10.0
openai==1.3.0
python-dateutil==2.8.2
gunicorn==21.2.0 